feat: init release page
