+++ client/components/HomeHeader.tsx
@@ -65,6 +65,12 @@ const HomeHeader = () => {
           {I18N.app.page.wenDang}
         </a>
         <a
+          href="/release"
+          className="mt-8 text-white lg:mt-0 lg:ml-8 lg:text-[#f4f4f5] lg:opacity-60 transition-opacity hover:opacity-90"
+        >
+          {I18N.app.page.release}
+        </a>
+        <a
           className="mt-8 text-white lg:hidden"
           onClick={() => setShowMobileNav(false)}
         >
